Extract DPM proof-pack supportability mapping
