+++ quality/architecture_rules.md
@@ -35,14 +35,18 @@ report-only baseline is reviewed.
 
 The largest current modularity risks are:
 
-1. `src/app/services/dpm_proof_pack_service.py` at 399 lines,
-2. `src/app/contracts/advisor_brief.py` at 398 lines,
-3. `src/app/services/advisor_brief_service.py` at 397 lines,
-4. `src/app/services/risk_workspace_attribution_controls.py` at 397 lines.
+1. `src/app/contracts/advisor_brief.py` at 398 lines,
+2. `src/app/services/advisor_brief_service.py` at 397 lines,
+3. `src/app/services/risk_workspace_attribution_controls.py` at 397 lines,
+4. `src/app/services/dpm_command_center_service.py` at 396 lines.
 
 `src/app/services/risk_workspace_service.py` is reduced from 402 to 222 lines after upstream
 summary, concentration, drawdown, rolling, and rolling-Sharpe fallback response loading moved into
 `src/app/services/risk_workspace_response_loading.py`.
+
+`src/app/services/dpm_proof_pack_service.py` is reduced from 399 to 315 lines after Manage
+proof-pack supportability mapping moved into
+`src/app/services/dpm_proof_pack_supportability.py`.
 
 `src/app/services/performance_workspace_contribution.py` is reduced from 402 to 227 lines after
 contribution level, row, position, smoothing-evidence, and source-economics payload mapping moved
