+++ src/app/services/dpm_proof_pack_service.py
@@ -14,6 +14,7 @@ from app.contracts.dpm_proof_packs import (
 )
 from app.services.ai_client_protocols import LotusAiWorkflowClient
 from app.services.dpm_client_protocols import DpmProofPackClient
+from app.services.dpm_proof_pack_supportability import build_dpm_proof_pack_supportability
 from app.services.lotus_ai_workflow import (
     build_workflow_pack_task_request,
     require_lotus_ai_client,
@@ -167,7 +168,7 @@ class DpmProofPackService:
         return ProofPackAiEvidenceInput(
             upstream_status=manage_status,
             payload=manage_payload,
-            supportability=_supportability_from(manage_payload),
+            supportability=build_dpm_proof_pack_supportability(manage_payload),
         )
 
     def _compose_response(
@@ -181,34 +182,11 @@ class DpmProofPackService:
             correlation_id=correlation_id,
             upstream_status=upstream_status,
             upstream_payload=upstream_payload,
-            supportability=_supportability_from(upstream_payload),
+            supportability=build_dpm_proof_pack_supportability(upstream_payload),
             error_model=DpmProofPackErrorDetail,
             error_code="MANAGE_PROOF_PACK_UPSTREAM_ERROR",
             default_detail="lotus-manage proof-pack request failed",
         )
-
-
-def _supportability_from(payload: dict[str, Any]) -> DpmProofPackSupportability:
-    proof_pack = _proof_pack_payload(payload)
-    reason_codes = _reason_codes_from_payload(payload, proof_pack)
-    return DpmProofPackSupportability(
-        state=_proof_pack_state(payload, proof_pack),
-        proof_pack_id=_safe_str(proof_pack.get("proof_pack_id") or payload.get("proof_pack_id")),
-        reason_codes=sorted(set(reason_codes)),
-        section_state_counts=_section_state_counts(proof_pack),
-        content_hash=_safe_str(proof_pack.get("content_hash") or payload.get("content_hash")),
-        markdown_available=bool(payload.get("markdown_url") or payload.get("markdown")),
-        report_input_available=bool(
-            payload.get("report_input_url")
-            or payload.get("report_input")
-            or proof_pack.get("report_input_ref")
-        ),
-        ai_evidence_input_available=bool(
-            payload.get("ai_evidence_input_url")
-            or payload.get("ai_evidence_input")
-            or proof_pack.get("ai_evidence_input_ref")
-        ),
-    )
 
 
 def _proof_pack_pm_memo_request_payload(
@@ -294,68 +272,6 @@ def _proof_pack_pm_memo_response(
         memo_request=memo_request,
         data=data,
     )
-
-
-def _proof_pack_payload(payload: dict[str, Any]) -> dict[str, Any]:
-    proof_pack = payload.get("proof_pack")
-    return proof_pack if isinstance(proof_pack, dict) else payload
-
-
-def _proof_pack_state(payload: dict[str, Any], proof_pack: dict[str, Any]) -> str:
-    state = (
-        proof_pack.get("status")
-        or proof_pack.get("state")
-        or payload.get("supportability_status")
-        or payload.get("supportabilityStatus")
-        or payload.get("status")
-        or "UNKNOWN"
-    )
-    return str(state)
-
-
-def _reason_codes_from_payload(
-    payload: dict[str, Any],
-    proof_pack: dict[str, Any],
-) -> list[str]:
-    reason_codes = _list_of_strings(payload.get("reason_codes") or payload.get("reasonCodes") or [])
-    reason_codes.extend(
-        _list_of_strings(proof_pack.get("reason_codes") or proof_pack.get("reasonCodes") or [])
-    )
-    for section in _sections(proof_pack):
-        reason_codes.extend(
-            _list_of_strings(section.get("reason_codes") or section.get("reasonCodes") or [])
-        )
-    return reason_codes
-
-
-def _section_state_counts(proof_pack: dict[str, Any]) -> dict[str, int]:
-    counts: dict[str, int] = {}
-    for section in _sections(proof_pack):
-        state = section.get("state") or section.get("status")
-        if state is None:
-            continue
-        state_label = str(state)
-        counts[state_label] = counts.get(state_label, 0) + 1
-    return counts
-
-
-def _sections(proof_pack: dict[str, Any]) -> list[dict[str, Any]]:
-    sections = proof_pack.get("sections")
-    if not isinstance(sections, list):
-        return []
-    return [section for section in sections if isinstance(section, dict)]
-
-
-def _list_of_strings(value: object) -> list[str]:
-    if not isinstance(value, list):
-        return []
-    return [str(item) for item in value]
-
-
-def _safe_str(value: object) -> str | None:
-    if value is None:
-        return None
-    return str(value)
 
 
 def _proof_pack_ai_source_refs(payload: dict[str, Any], proof_pack_id: str) -> list[str]:
